refactor - move java source files to java src folder
