+++ src/main/java/aoc/day1/Main.java
@@ -1,4 +1,4 @@
-package day1;
+package aoc.day1;
 
 import io.ReadFile;
