+++ src/main/java/aoc/day2/Main.java
@@ -1,4 +1,4 @@
-package day2;
+package aoc.day2;
 
 import io.ReadFile;
