+++ src/main/java/aoc/day2/games.txt
