+++ src/main/java/aoc/day3/Main.java
@@ -1,4 +1,4 @@
-package day3;
+package aoc.day3;
 
 import io.ReadFile;
