+++ src/main/java/aoc/day3/SumPartNumbers.java
@@ -1,4 +1,4 @@
-package day3;
+package aoc.day3;
 
 public class SumPartNumbers {
     public int sumPartNumbers (String filename) {
