+++ src/main/java/aoc/day3/engine_schematic.txt
