+++ src/main/java/aoc/day4/Main.java
@@ -1,4 +1,4 @@
-package day4;
+package aoc.day4;
 
 import io.ReadFile;
