+++ src/main/java/aoc/day4/TotalCardNumber.java
@@ -1,4 +1,4 @@
-package day4;
+package aoc.day4;
 
 import java.util.stream.*;
