+++ src/main/java/aoc/day4/scratchcard.txt
