+++ src/main/java/aoc/io/ReadFile.java
@@ -1,4 +1,4 @@
-package io;
+package aoc.io;
 
 import java.io.File;
 import java.io.FileNotFoundException;
